requirements.txt
opencv-python
pillow
tesseract
